+++ packages/girder-redux/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.23](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/girder-redux@0.0.22...@openchemistry/girder-redux@0.0.23) (2019-07-24)
+
+**Note:** Version bump only for package @openchemistry/girder-redux
+
+
+
+
+
 ## [0.0.22](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/girder-redux@0.0.21...@openchemistry/girder-redux@0.0.22) (2019-06-03)
 
 **Note:** Version bump only for package @openchemistry/girder-redux