+++ packages/molecule/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.22](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/molecule@0.4.21...@openchemistry/molecule@0.4.22) (2019-07-24)
+
+**Note:** Version bump only for package @openchemistry/molecule
+
+
+
+
+
+## [0.4.21](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/molecule@0.4.20...@openchemistry/molecule@0.4.21) (2019-07-09)
+
+**Note:** Version bump only for package @openchemistry/molecule
+
+
+
+
+
+## [0.4.20](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/molecule@0.4.19...@openchemistry/molecule@0.4.20) (2019-06-29)
+
+**Note:** Version bump only for package @openchemistry/molecule
+
+
+
+
+
+## [0.4.19](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/molecule@0.4.18...@openchemistry/molecule@0.4.19) (2019-06-27)
+
+**Note:** Version bump only for package @openchemistry/molecule
+
+
+
+
+
 ## [0.4.18](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/molecule@0.4.17...@openchemistry/molecule@0.4.18) (2019-05-24)
 
 **Note:** Version bump only for package @openchemistry/molecule