+++ packages/sagas/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.33](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/sagas@0.3.32...@openchemistry/sagas@0.3.33) (2019-07-24)
+
+**Note:** Version bump only for package @openchemistry/sagas
+
+
+
+
+
+## [0.3.32](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/sagas@0.3.31...@openchemistry/sagas@0.3.32) (2019-06-27)
+
+**Note:** Version bump only for package @openchemistry/sagas
+
+
+
+
+
 ## [0.3.31](https://github.com/OpenChemistry/oc-web-components/compare/@openchemistry/sagas@0.3.30...@openchemistry/sagas@0.3.31) (2019-06-03)
 
 **Note:** Version bump only for package @openchemistry/sagas